etc: 디렉토리 구조 변경
- exception 디렉토리 별도로 분리
- DTO는 request와 response 디렉토리 별도로 분리
