+++ src/main/java/hyeong/lee/myboard/dto/request/ReplyRequestDto.java
@@ -1,39 +1,34 @@
-package hyeong.lee.myboard.dto;
+package hyeong.lee.myboard.dto.request;
 
 import hyeong.lee.myboard.domain.Board;
+import hyeong.lee.myboard.domain.Reply;
 import hyeong.lee.myboard.domain.UserAccount;
 import lombok.*;
-import org.hibernate.validator.constraints.Range;
-import org.springframework.web.multipart.MultipartFile;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
 
-@ToString
 @Getter @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class BoardRequestDto {
-
-    @Size(max = 50)
-    @NotBlank
-    private String title;
+public class ReplyRequestDto {
+        
+    private Long boardId;
 
     @Size(max = 15)
     @NotBlank
     private String editor;
-
+    
     @NotBlank
     private String content;
 
-    private MultipartFile[] files;
     private UserAccountDto userAccountDto;
 
-    public Board toEntity() {
+    public Reply toEntity(Board board) {
         UserAccount userAccount = userAccountDto == null ? null : userAccountDto.toEntity();
 
-        return Board.builder()
-                .title(title)
+        return Reply.builder()
+                .board(board)
                 .editor(editor)
                 .content(content)
                 .userAccount(userAccount)
